EventIndex: Rename the file to be consistent with the class.
